project uploaded
In this project, we created a Vehicle Management System (VMS) for individuals seeking car repair services. Through our website and app, users can conveniently access a list of nearby garages along with their pricing information, making it easier to book and receive quality service. The technology stack employed includes Spring Boot and Java for the business layer, Bootstrap for the framework, HTML, CSS, and JavaScript for the presentation layer, and MySQL for the database.